fixed modulation
set modulation does set snmp call then assigns the value based on the
get of that same oid. This is done because the set function returns the
value you sent instead of the value that got set. i.e. setting a bad
value still returns the bad value instead of the value the OID has.